+++ normalized-cache/src/commonMain/kotlin/com/apollographql/cache/normalized/api/ReadOnlyNormalizedCache.kt
@@ -25,9 +25,24 @@ interface ReadOnlyNormalizedCache {
 
   /**
    * Returns the size in bytes of a [Record].
-   * This is an optional operation that can be implemented by the caches for debug purposes, otherwise it defaults to `-1`, meaning unknown size.
+   * This is an optional operation that can be implemented by caches for debug purposes, otherwise it defaults to `-1`, meaning unknown size.
    */
   fun sizeOfRecord(record: Record): Int = -1
+
+  /**
+   * Returns the total size in bytes of this cache.
+   * This is an optional operation that can be implemented by caches for debug or observability purposes, otherwise it defaults to `-1`, meaning unknown size.
+   *
+   * Note: in case of a chained cache, this size does not include the size of the next cache in the chain. See [nextCache].
+   */
+  suspend fun size(): Long = -1L
+
+  /**
+   * Some cache implementations (e.g. [com.apollographql.cache.normalized.internal.OptimisticNormalizedCache] and [com.apollographql.cache.normalized.memory.MemoryCache]) support chaining caches.
+   * For those, this returns the next cache in the chain or `null` if there is none. Implementations that don't support chaining return `null`.
+   */
+  val nextCache: ReadOnlyNormalizedCache?
+    get() = null
 
   suspend fun close() {}
 }