+++ tests/pagination/src/commonTest/kotlin/ConnectionPaginationTest.kt
@@ -2,6 +2,7 @@ package pagination
 
 import com.apollographql.apollo.api.Error
 import com.apollographql.apollo.api.Optional
+import com.apollographql.apollo.exception.apolloExceptionHandler
 import com.apollographql.cache.normalized.CacheManager
 import com.apollographql.cache.normalized.api.ConnectionFieldKeyGenerator
 import com.apollographql.cache.normalized.api.ConnectionMetadataGenerator
@@ -305,6 +306,7 @@ class ConnectionPaginationTest {
     assertChainedCachesAreEqual(cacheManager)
 
     // Non-contiguous page (should be ignored)
+    apolloExceptionHandler = {}
     val query5 = UsersQuery(first = Optional.Present(2), after = Optional.Present("xx50"))
     val data5 = UsersQuery.Data {
       users = buildUserConnection {